Split reducer specs
