+++ react-app/__tests__/reducers/todos-test.js
@@ -1,4 +1,3 @@
-// __tests__/TodoApp.react-test.js
 jest.dontMock('../../js/reducers/todoApp');
 jest.dontMock('../../js/reducers/authentication');
 jest.dontMock('../../js/reducers/todos');
@@ -58,50 +57,6 @@ describe('todos FETCH_TASKS_SUCCESS', () => {
     expect(newState.data).toEqual({
       tasks: tasks,
       newTask: { id: 0, title: '', completed: false }
-    });
-  });
-});
-
-describe('todoApp LOGIN_SUCCESS', () => {
-  it('sets the user and login state', () => {
-    let newState = todoApp(undefined, {
-      type: 'LOGIN_SUCCESS',
-      user: { id: 123, name: 'Bob' }
-    });
-    expect(newState).toEqual({
-      authentication: {
-        loggedIn: true,
-        user: { id: 123, name: 'Bob' },
-      },
-      data: {
-        tasks: [],
-        newTask: { id: 0, title: '', completed: false }
-      }
-    });
-  });
-});
-
-describe('todoApp LOGIN_FAILURE', () => {
-  it('sets the user and login state', () => {
-    let previousState = Object.assign({}, todoApp.INITIAL_STATE, {
-      authentication: {
-        loggedIn: true,
-        user: { id: 456, name: 'Alice' }
-      }
-    });
-    let newState = todoApp(previousState, {
-      type: 'LOGIN_FAILURE',
-      error: 'You are not getting in'
-    });
-    expect(newState).toEqual({
-      authentication: {
-        loggedIn: false,
-        user: null,
-      },
-      data: {
-        tasks: [],
-        newTask: { id: 0, title: '', completed: false }
-      }
     });
   });
 });
